refactor: mv src/gui/src/ActionDialog => src/gui/src/dialogs/ActionDialog
